day3

+++ S4/Day3/backend/routes/note.route.js
@@ -1,11 +1,17 @@
 const express = require("express")
 const noteRouter = express.Router()
 const {NoteModel}=require("../model/note.model")
+const jwt = require("jsonwebtoken")
 
 noteRouter.get("/",async (req,res)=>{
+   const token = req.headers.authorization.split(" ")[1];
+   const decoded = jwt.verify(token,"bruce")
  try {
-    const note = await NoteModel.find()
-    res.status(200).send(note)
+   if(decoded){
+      const note = await NoteModel.find({"userID":decoded.userID})
+      res.status(200).send(note)
+   }
+    
  } catch (err) {
     res.status(400).send({"msg":err.message})
  }
@@ -37,12 +43,21 @@ try {
 
 noteRouter.delete("/delete/:noteID",async(req,res)=>{
 //logic
-     const payload = req.body
+     const token = req.headers.authorization.split(" ")[1]
+     const decoded = jwt.verify(token,"bruce")
+     const req_id = decoded.userID
+     const note = NoteModel.findOne({_id:noteID})
+     const userID_in_note = note.userID
+     //const payload = req.body
      const noteID = req.params.noteID
-     try {
-        await NoteModel.findByIdAndDelete({_id:noteID})
-        res.status(200).send({"msg":"Notes has been Deleted"})
 
+     try {
+      if(req_id===userID_in_note){
+         await NoteModel.findByIdAndDelete({_id:noteID})
+        res.status(200).send({"msg":"Notes has been Deleted"})
+      }else{
+         res.status(400).send({"msg":"Not Authorised"})
+      }
      } catch (err) {
         res.status(400).send({"msg":err.message}) 
 
@@ -53,3 +68,5 @@ noteRouter.delete("/delete/:noteID",async(req,res)=>{
 module.exports = {
     noteRouter
 }
+
+
